+++ src/dict/many/tree.ts
@@ -1,20 +1,35 @@
 
-import MultiTreeDict from "../multi/tree"
+// import MultiTreeDict from "../multi/tree"
+import AVL from "../../avl"
 import { Cursor } from "../../interfaces"
 import { lesser, equal } from "../../util"
+import { MultiDict } from "../../interfaces"
+import { TreeDictOptions } from "../tree"
 
-export class TreeDict<K, V> extends MultiTreeDict<K, V> {
+export class TreeManyDict<K, V> extends AVL<[K, V], K> implements MultiDict<K, V> {
 
-  constructor(
-      keyLessThan: (a: K, b: K) => boolean = lesser
-    , valuesEqual: (a: V, b: V) => boolean = equal
-  ) {
-    super(keyLessThan, valuesEqual);
+  static empty<K, V>(opts: TreeDictOptions<K, V> = {}) {
+    const valuesEqual = opts.valuesEqual !== undefined ? opts.valuesEqual : equal;
+    return new TreeManyDict<K, V>(
+      opts.compare !== undefined ? opts.compare : lesser
+    , pair => pair[0]
+    , (a, b) => valuesEqual(a[1], b[1])
+    , true
+    );
+  }
+
+  static from<K, V>(iterable: Iterable<[K, V]>, opts?: TreeDictOptions<K, V>) {
+    // FIXME might be able to optimise this
+    const dict = TreeManyDict.empty(opts);
+    for (const element of iterable) {
+      dict.add(element);
+    }
+    return dict;
   }
 
   add(pair: [K, V]): [boolean, Cursor<[K, V]>] {
     for (const node of this.equalKeys(pair[0])) {
-      if (this.valuesEqual(pair[1], node.value[1])) {
+      if (this.elementsEqual(pair, node.value)) {
         node.value = pair;
         return [false, node];
       }
@@ -22,7 +37,26 @@ export class TreeDict<K, V> extends MultiTreeDict<K, V> {
     return super.add(pair) as [boolean, Cursor<[K, V]>];
   }
 
+  *getValues(key: K) {
+    for (const cursor of this.equalKeys(key)) {
+      yield cursor.value[1];
+    }
+  }
+
+  emplace(key: K, val: V) {
+    return this.add([key, val]);
+  }
+
+  clone() {
+    return new TreeManyDict<K, V>(
+      this.lessThan
+    , pair => pair[0]
+    , this.elementsEqual
+    , true
+    );
+  }
+
 }
 
-export default TreeDict;
+export default TreeManyDict;