+++ src/test/_helpers.ts
@@ -18,74 +18,62 @@ const TEST_FILES = [
   {
     name: 'TreeMultiDict',
     file: '../dict/multi/tree',
-    implements: ['MultiDict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new TreeMultiDict<any, any>(...args),
+    implements: ['MultiDict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'HashMultiDict',
     file: '../dict/multi/hash',
-    implements: ['MultiDict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new HashMultiDict<any, any>(...args),
+    implements: ['MultiDict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'TreeManyDict',
     file: '../dict/many/tree',
-    implements: ['ManyDict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new TreeManyDict<any, any>(...args),
+    implements: ['ManyDict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'HashManyDict',
     file: '../dict/many/hash',
-    implements: ['ManyDict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new HashManyDict<any, any>(...args),
+    implements: ['ManyDict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'TreeDict',
     file: '../dict/tree',
-    implements: ['Dict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new TreeDict<any, any>(...args),
+    implements: ['Dict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'HashDict',
     file: '../dict/hash',
-    implements: ['Dict', 'DictLike', 'KeyedCollection', 'Collection'],
-    create: (...args: any[]) => new HashDict<any, any>(...args),
+    implements: ['Dict', 'DictLike', 'IndexedCollection', 'CollectionLike'],
   },
   {
     name: 'Queue',
     file: '../queue',
     implements: ['Queuelike', 'Collection'],
-    create: (...args: any[]) => new Queue<any>(...args),
   },
   {
     name: 'Stack',
     file: '../stack',
     implements: ['Queuelike', 'Collection'],
-    create: (...args: any[]) => new Stack<any>(...args),
   },
   {
     name: 'PriorityQueue',
     file: '../priority-queue',
     implements: ['Queuelike', 'Collection'],
-    create: (...args: any[]) => new PriorityQueue<any>(...args),
   },
   {
     name: 'DoubleLinkedList',
     file: '../list/double',
     implements: ['Sequence', 'List', 'Collection'],
-    create: (...args: any[]) => new DoubleLinkedList<any>(...args),
   },
   {
     name: 'SingleLinkedList',
     file: "../list/single",
     implements: ['Sequence', 'List', 'Collection'],
-    create: (...args: any[]) => new SingleLinkedList<any>(...args),
   },
   {
     name: 'Vector',
     file: '../vector',
     implements: ['Sequence', 'Collection'],
-    create: (...args: any[]) => new Vector<any>(...args),
   },
 ]
 
@@ -118,7 +106,7 @@ export function test<C extends Collection<any>>(name: string, callback: (collect
       describe(testFile.name + methodName, () => {
         it(description, () => {
           const args = (opts && opts.args) || [];
-          callback(testFile.create(...args) as any);
+          callback(require(testFile.file).default.empty(...args) as any);
         })
       })
     }